Acertos finais dos Desafios

+++ Desafio7.2.py
@@ -2,12 +2,12 @@
 import statistics
 
 def receber_notas():
-    n1 = int(input("Insira a primeira nota: "))
-    n2 = int(input("Insira a segunda nota: "))
+    n1 = float(input("Insira a primeira nota: "))
+    n2 = float(input("Insira a segunda nota: "))
     return n1, n2
 
 def average(receber_notas):
-    soma = int(receber_notas[0] + receber_notas[1])
+    soma = float(receber_notas[0] + receber_notas[1])
     media = soma / 2
     return print(media)
